base_quantumblur/requirements.txt
Pillow
wheel
qiskit-terra
qiskit-aer